fix(tests): strengthen assertions, fix spec drift, split oversized files (#21)
- Use real stub tokenizer + assert exact fallback format in mlx test
- Strengthen assertions, fix spec drift, split oversized files
