+++ tests/test_daemon.py
@@ -100,58 +100,52 @@ class TestHandleRequest:
         assert response.endswith(b"\n")
 
 
+def _send_request(sock_path: str, payload: dict) -> dict:
+    encoded = json.dumps(payload).encode() + b"\n"
+    with socket.socket(socket.AF_UNIX, socket.SOCK_STREAM) as sock:
+        sock.settimeout(3.0)
+        sock.connect(sock_path)
+        sock.sendall(encoded)
+        data = b""
+        while True:
+            chunk = sock.recv(4096)
+            if not chunk or b"\n" in data + chunk:
+                data += chunk
+                break
+            data += chunk
+    return json.loads(data.decode().strip())
+
+
 class TestDaemonSocketProtocol:
     def test_daemon_serves_request_via_socket(self) -> None:
+        import os
         import tempfile
 
-        # Unix sockets have a 104-char path limit on macOS — use /tmp directly
         with tempfile.NamedTemporaryFile(suffix=".sock", dir="/tmp", delete=False) as f:
             socket_path = f.name
         with tempfile.NamedTemporaryFile(suffix=".pid", dir="/tmp", delete=False) as f:
             pid_file = f.name
-        import os
-
-        os.unlink(socket_path)  # daemon will re-create it
+        os.unlink(socket_path)
         backend = MockBackend(verdict="clean", reason="socket test")
-
-        import os
 
         plugin_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
         daemon = VaudevilleDaemon(socket_path, pid_file, plugin_root, backend)
 
         thread = threading.Thread(target=daemon.serve, daemon=True)
         thread.start()
-        time.sleep(0.2)  # Brief pause for socket to bind
+        time.sleep(0.2)
 
-        with socket.socket(socket.AF_UNIX, socket.SOCK_STREAM) as sock:
-            sock.settimeout(3.0)
-            sock.connect(socket_path)
-            payload = (
-                json.dumps(
-                    {
-                        "rule": "violation-detector",
-                        "input": {"text": "All good."},
-                    }
-                ).encode()
-                + b"\n"
-            )
-            sock.sendall(payload)
-            data = b""
-            while True:
-                chunk = sock.recv(4096)
-                if not chunk or b"\n" in data + chunk:
-                    data += chunk
-                    break
-                data += chunk
-
-        response = json.loads(data.decode().strip())
+        response = _send_request(
+            socket_path,
+            {"rule": "violation-detector", "input": {"text": "All good."}},
+        )
         assert response["verdict"] == "clean"
         daemon._stop_event.set()
 
     def test_oversized_payload_returns_fail_open(self) -> None:
-        """Daemon drops payloads exceeding MAX_REQUEST_SIZE and fails open."""
-        import tempfile
         import os
+        import tempfile
+
         from vaudeville.server.daemon import MAX_REQUEST_SIZE
 
         with tempfile.NamedTemporaryFile(suffix=".sock", dir="/tmp", delete=False) as f:
@@ -171,7 +165,6 @@ class TestDaemonSocketProtocol:
         with socket.socket(socket.AF_UNIX, socket.SOCK_STREAM) as sock:
             sock.settimeout(3.0)
             sock.connect(socket_path)
-            # Send payload larger than MAX_REQUEST_SIZE with no newline
             sock.sendall(b"x" * (MAX_REQUEST_SIZE + 1))
             data = b""
             while True:
@@ -184,7 +177,7 @@ class TestDaemonSocketProtocol:
 
         response = json.loads(data.decode().strip())
         assert response["verdict"] == "clean"
-        assert backend.calls == []  # backend should never be called
+        assert backend.calls == []
         daemon._stop_event.set()
